Start commandline quoting
do not use StringUtil
