+++ jetty-core/jetty-start/pom.xml
@@ -70,7 +70,6 @@
                     <include>org/eclipse/jetty/util/ManifestUtils*</include>
                     <include>org/eclipse/jetty/util/TopologicalSort*</include>
                     <include>org/eclipse/jetty/util/FileID*</include>
-                    <include>org/eclipse/jetty/util/StringUtil*</include>
                   </includes>
                 </filter>
               </filters>
