+++ jetty-core/jetty-start/src/test/java/org/eclipse/jetty/start/CommandLineBuilderTest.java
@@ -13,7 +13,12 @@
 
 package org.eclipse.jetty.start;
 
+import java.util.stream.Stream;
+
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -66,6 +71,25 @@ public class CommandLineBuilderTest
         cmd.addArg("-Djetty.home", "/opt/jetty");
         cmd.addArg("monetary.symbol", "€");
         assertThat(cmd.toCommandLine(), is("java -Djetty.home=/opt/jetty monetary.symbol=€"));
+    }
+
+    public static Stream<Arguments> shellQuoting()
+    {
+        return Stream.of(
+            Arguments.of(null, null),
+            Arguments.of("", ""),
+            Arguments.of("Hello", "Hello"),
+            Arguments.of("Hello World", "\"Hello World\""),
+            Arguments.of("foo\\bar", "\"foo\\\\bar\""),
+            Arguments.of("'single'", "\"'single'\"")
+        );
+    }
+
+    @ParameterizedTest
+    @MethodSource("shellQuoting")
+    public void testShellQuoting(String string, String expected)
+    {
+        assertThat(CommandLineBuilder.shellQuoteIfNeeded(string), is(expected));
     }
 
     @Test
